Initialized internal/cli pkg #95
Fixed reference err in client factory func and test initial cli layout

WIP: test client intermission

Simplified types and kinds for messages and events

Simplified client/relay #listenConn

Improve typing

Use #Reader vs #Read in client/relay

Improved typing #130

Simplifed types with aliases

Simplified tags with aliases

Clean-up add back interfaces

Optimize types further

Optimize types further

Fixed type err in AmountTag

Mucho improvmento

+++ count.go
@@ -2,5 +2,5 @@ package nostr
 
 // Count TBD
 type Count struct {
-	Count float64 `json:"count,omitempty"`
+	Count int `json:"count,omitempty"`
 }
